Add client.nextPage to get next page of collection
