+++ lib/client.js
@@ -4,6 +4,7 @@ var BasicAuthenticator = require('./auth/basic_authenticator');
 var OauthAuthenticator = require('./auth/oauth_authenticator');
 var App = require('./auth/app');
 var autoDetect = require('./auth/auto_detect');
+var paging = require('./util/paging.js');
 
 /**
  * Constructs a Client with instances of all the resources using the dispatcher.
@@ -133,6 +134,17 @@ Client.prototype.useOauth = function(options) {
   }
   this.dispatcher.setAuthenticator(authenticator);
   return this;
+};
+
+/**
+ * Get the next page of results in a collection.
+ *
+ * @param {Object} response Full payload from a response to a
+ *     collection request.
+ * @param {Object} [dispatchOptions]
+ */
+Client.prototype.nextPage = function(response, dispatchOptions) {
+  return paging.nextPage(response, this.dispatcher, dispatchOptions);
 };
 
 /**
